+++ package.json
@@ -1,18 +1,21 @@
 {
   "name": "yt-music-speed-normalizer",
-  "version": "1.0.0",
-  "main": "index.js",
-  "author": "https://github.com/hamao0820",
-  "license": "MIT",
+  "module": "index.ts",
+  "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc --noEmit && vite build",
-    "preview": "vite preview"
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "zip:firefox": "wxt zip -b firefox",
+    "postinstall": "wxt prepare"
   },
   "devDependencies": {
-    "@crxjs/vite-plugin": "^1.0.14",
-    "@types/chrome": "^0.0.271",
-    "typescript": "^5.6.2",
-    "vite": "^2.9.4"
+    "@types/bun": "latest",
+    "wxt": "^0.19.13"
+  },
+  "peerDependencies": {
+    "typescript": "^5.0.0"
   }
 }
