--- release.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-yarn build
-mkdir -p release
-zip -r release/release.zip dist/
