--- src/type.d.ts
@@ -1,3 +1,0 @@
-export type ChromeMessage = {
-  type: string;
-};
